+++ skan.potal.com/src/main/java/com/skan/potal/web/potal/cattle/dto/HmCattleBuyInfo.java
@@ -1,10 +1,9 @@
-package com.skan.potal.web.potal.cattle.model;
+package com.skan.potal.web.potal.cattle.dto;
 
 import java.io.Serializable;
 import java.util.Date;
 
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.IdClass;
 import javax.persistence.JoinColumn;
@@ -12,118 +11,129 @@ import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 /**
- * 판매지정보 모델 클래스.
+ * 구입기록 모델 클래스.
  * 
  * @author generated by ERMaster
  * @version $Id$
  */
 @Entity
 @Table
-public class HmCattleSellStoreInfo implements Serializable {
+@IdClass(HmCattlePK.class)
+public class HmCattleBuyInfo implements Serializable {
 
 	/** serialVersionUID. */
 	private static final long serialVersionUID = 1L;
-  
-	/** 판매처. */
-	private String storeName;
 
-	/** 판매/도태일자. */
-	private Date sellDate;
+	/** 구 입 처. */
+	private String buyStoreName;
 
-	/** 비고. */
-	private String sellNote;
+	/** 구 입 일 자. */
+	private Date buyDate;
 
-	/** 판매 연락처. */
-	private String sellPhoneNumber;
-	
+	/** 비 고. */
+	private String buyNote;
+
+	/** 구입 연락처. */
+	private String buyPhoneNumber;
+
 	@Id
 	private String entityDiscernNo;
-
+	
 	/** 한우(암소/수소) 개체관리기록부. */
-	@ManyToOne(fetch=FetchType.EAGER)
-	@JoinColumn(name="entityDiscernNo" , insertable=false, updatable=false)
+	@ManyToOne
+	@JoinColumn(name="entityDiscernNo" ,insertable = false, updatable = false, referencedColumnName="entityDiscernNo")
 	private HmCattleRegister hmCattleRegister;
 
 	/**
 	 * 생성자.
 	 */
-	public HmCattleSellStoreInfo() {
+	public HmCattleBuyInfo() {
+	}
+	
+	public String getEntityDiscernNo() {
+		return entityDiscernNo;
+	}
+
+	public void setEntityDiscernNo(String entityDiscernNo) {
+		this.entityDiscernNo = entityDiscernNo;
+	}
+
+
+
+	/**
+	 * 구 입 처을 설정합니다..
+	 * 
+	 * @param buyStoreName
+	 *            구 입 처
+	 */
+	public void setBuyStoreName(String buyStoreName) {
+		this.buyStoreName = buyStoreName;
 	}
 
 	/**
-	 * 판매처을 설정합니다..
+	 * 구 입 처을 가져옵니다..
 	 * 
-	 * @param storeName
-	 *            판매처
+	 * @return 구 입 처
 	 */
-	public void setStoreName(String storeName) {
-		this.storeName = storeName;
+	public String getBuyStoreName() {
+		return this.buyStoreName;
 	}
 
 	/**
-	 * 판매처을 가져옵니다..
+	 * 구 입 일 자을 설정합니다..
 	 * 
-	 * @return 판매처
+	 * @param buyDate
+	 *            구 입 일 자
 	 */
-	public String getStoreName() {
-		return this.storeName;
+	public void setBuyDate(Date buyDate) {
+		this.buyDate = buyDate;
 	}
 
 	/**
-	 * 판매/도태일자을 설정합니다..
+	 * 구 입 일 자을 가져옵니다..
 	 * 
-	 * @param sellDate
-	 *            판매/도태일자
+	 * @return 구 입 일 자
 	 */
-	public void setSellDate(Date sellDate) {
-		this.sellDate = sellDate;
+	public Date getBuyDate() {
+		return this.buyDate;
 	}
 
 	/**
-	 * 판매/도태일자을 가져옵니다..
+	 * 비 고을 설정합니다..
 	 * 
-	 * @return 판매/도태일자
+	 * @param buyNote
+	 *            비 고
 	 */
-	public Date getSellDate() {
-		return this.sellDate;
+	public void setBuyNote(String buyNote) {
+		this.buyNote = buyNote;
 	}
 
 	/**
-	 * 비고을 설정합니다..
+	 * 비 고을 가져옵니다..
 	 * 
-	 * @param sellNote
-	 *            비고
+	 * @return 비 고
 	 */
-	public void setSellNote(String sellNote) {
-		this.sellNote = sellNote;
+	public String getBuyNote() {
+		return this.buyNote;
 	}
 
 	/**
-	 * 비고을 가져옵니다..
+	 * 구입 연락처을 설정합니다..
 	 * 
-	 * @return 비고
+	 * @param buyPhoneNumber
+	 *            구입 연락처
 	 */
-	public String getSellNote() {
-		return this.sellNote;
+	public void setBuyPhoneNumber(String buyPhoneNumber) {
+		this.buyPhoneNumber = buyPhoneNumber;
 	}
 
 	/**
-	 * 판매 연락처을 설정합니다..
+	 * 구입 연락처을 가져옵니다..
 	 * 
-	 * @param sellPhoneNumber
-	 *            판매 연락처
+	 * @return 구입 연락처
 	 */
-	public void setSellPhoneNumber(String sellPhoneNumber) {
-		this.sellPhoneNumber = sellPhoneNumber;
-	}
-
-	/**
-	 * 판매 연락처을 가져옵니다..
-	 * 
-	 * @return 판매 연락처
-	 */
-	public String getSellPhoneNumber() {
-		return this.sellPhoneNumber;
+	public String getBuyPhoneNumber() {
+		return this.buyPhoneNumber;
 	}
 
 	/**
@@ -143,14 +153,6 @@ public class HmCattleSellStoreInfo implements Serializable {
 	 */
 	public HmCattleRegister getHmCattleRegister() {
 		return this.hmCattleRegister;
-	}
-
-	public String getEntityDiscernNo() {
-		return entityDiscernNo;
-	}
-
-	public void setEntityDiscernNo(String entityDiscernNo) {
-		this.entityDiscernNo = entityDiscernNo;
 	}
 
 	/**
@@ -178,7 +180,7 @@ public class HmCattleSellStoreInfo implements Serializable {
 		if (getClass() != obj.getClass()) {
 			return false;
 		}
-		HmCattleSellStoreInfo other = (HmCattleSellStoreInfo) obj;
+		HmCattleBuyInfo other = (HmCattleBuyInfo) obj;
 		if (hmCattleRegister == null) {
 			if (other.hmCattleRegister != null) {
 				return false;
