+++ skan.potal.com/src/main/java/com/skan/potal/web/potal/cattle/dto/HmCattleCalfRecodeId.java
@@ -1,4 +1,4 @@
-package com.skan.potal.web.potal.cattle.model;
+package com.skan.potal.web.potal.cattle.dto;
 
 import java.io.Serializable;
 
@@ -7,13 +7,13 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 /**
- * 질병 및 치료 composite - id 클래스.
+ * 송아지 기록 composite - id 클래스.
  * 
  * @author generated by ERMaster
  * @version $Id$
  */
 @Embeddable
-public class HmCattleCureInfoId implements Serializable {
+public class HmCattleCalfRecodeId implements Serializable {
 
 	/** serialVersionUID. */
 	private static final long serialVersionUID = 1L;
@@ -29,7 +29,7 @@ public class HmCattleCureInfoId implements Serializable {
 	/**
 	 * 생성자.
 	 */
-	public HmCattleCureInfoId() {
+	public HmCattleCalfRecodeId() {
 	}
 
 	/**
@@ -96,7 +96,7 @@ public class HmCattleCureInfoId implements Serializable {
 		if (getClass() != obj.getClass()) {
 			return false;
 		}
-		HmCattleCureInfoId other = (HmCattleCureInfoId) obj;
+		HmCattleCalfRecodeId other = (HmCattleCalfRecodeId) obj;
 		if (hmCattleRegister == null) {
 			if (other.hmCattleRegister != null) {
 				return false;
