+++ skan.potal.com/src/main/java/com/skan/potal/web/potal/cattle/dto/HmCattleCureInfoId.java
@@ -1,4 +1,4 @@
-package com.skan.potal.web.potal.cattle.model;
+package com.skan.potal.web.potal.cattle.dto;
 
 import java.io.Serializable;
 
@@ -7,13 +7,13 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 /**
- * 송아지 기록 composite - id 클래스.
+ * 질병 및 치료 composite - id 클래스.
  * 
  * @author generated by ERMaster
  * @version $Id$
  */
 @Embeddable
-public class HmCattleCalfRecodeId implements Serializable {
+public class HmCattleCureInfoId implements Serializable {
 
 	/** serialVersionUID. */
 	private static final long serialVersionUID = 1L;
@@ -29,7 +29,7 @@ public class HmCattleCalfRecodeId implements Serializable {
 	/**
 	 * 생성자.
 	 */
-	public HmCattleCalfRecodeId() {
+	public HmCattleCureInfoId() {
 	}
 
 	/**
@@ -96,7 +96,7 @@ public class HmCattleCalfRecodeId implements Serializable {
 		if (getClass() != obj.getClass()) {
 			return false;
 		}
-		HmCattleCalfRecodeId other = (HmCattleCalfRecodeId) obj;
+		HmCattleCureInfoId other = (HmCattleCureInfoId) obj;
 		if (hmCattleRegister == null) {
 			if (other.hmCattleRegister != null) {
 				return false;
