+++ skan.potal.com/src/main/java/com/skan/potal/web/potal/cattle/dto/HmCattleSellStoreInfo.java
@@ -1,9 +1,10 @@
-package com.skan.potal.web.potal.cattle.model;
+package com.skan.potal.web.potal.cattle.dto;
 
 import java.io.Serializable;
 import java.util.Date;
 
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.IdClass;
 import javax.persistence.JoinColumn;
@@ -11,129 +12,118 @@ import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 /**
- * 구입기록 모델 클래스.
+ * 판매지정보 모델 클래스.
  * 
  * @author generated by ERMaster
  * @version $Id$
  */
 @Entity
 @Table
-@IdClass(HmCattlePK.class)
-public class HmCattleBuyInfo implements Serializable {
+public class HmCattleSellStoreInfo implements Serializable {
 
 	/** serialVersionUID. */
 	private static final long serialVersionUID = 1L;
+  
+	/** 판매처. */
+	private String storeName;
 
-	/** 구 입 처. */
-	private String buyStoreName;
+	/** 판매/도태일자. */
+	private Date sellDate;
 
-	/** 구 입 일 자. */
-	private Date buyDate;
+	/** 비고. */
+	private String sellNote;
 
-	/** 비 고. */
-	private String buyNote;
-
-	/** 구입 연락처. */
-	private String buyPhoneNumber;
-
+	/** 판매 연락처. */
+	private String sellPhoneNumber;
+	
 	@Id
 	private String entityDiscernNo;
-	
+
 	/** 한우(암소/수소) 개체관리기록부. */
-	@ManyToOne
-	@JoinColumn(name="entityDiscernNo" ,insertable = false, updatable = false, referencedColumnName="entityDiscernNo")
+	@ManyToOne(fetch=FetchType.EAGER)
+	@JoinColumn(name="entityDiscernNo" , insertable=false, updatable=false)
 	private HmCattleRegister hmCattleRegister;
 
 	/**
 	 * 생성자.
 	 */
-	public HmCattleBuyInfo() {
-	}
-	
-	public String getEntityDiscernNo() {
-		return entityDiscernNo;
-	}
-
-	public void setEntityDiscernNo(String entityDiscernNo) {
-		this.entityDiscernNo = entityDiscernNo;
-	}
-
-
-
-	/**
-	 * 구 입 처을 설정합니다..
-	 * 
-	 * @param buyStoreName
-	 *            구 입 처
-	 */
-	public void setBuyStoreName(String buyStoreName) {
-		this.buyStoreName = buyStoreName;
+	public HmCattleSellStoreInfo() {
 	}
 
 	/**
-	 * 구 입 처을 가져옵니다..
+	 * 판매처을 설정합니다..
 	 * 
-	 * @return 구 입 처
+	 * @param storeName
+	 *            판매처
 	 */
-	public String getBuyStoreName() {
-		return this.buyStoreName;
+	public void setStoreName(String storeName) {
+		this.storeName = storeName;
 	}
 
 	/**
-	 * 구 입 일 자을 설정합니다..
+	 * 판매처을 가져옵니다..
 	 * 
-	 * @param buyDate
-	 *            구 입 일 자
+	 * @return 판매처
 	 */
-	public void setBuyDate(Date buyDate) {
-		this.buyDate = buyDate;
+	public String getStoreName() {
+		return this.storeName;
 	}
 
 	/**
-	 * 구 입 일 자을 가져옵니다..
+	 * 판매/도태일자을 설정합니다..
 	 * 
-	 * @return 구 입 일 자
+	 * @param sellDate
+	 *            판매/도태일자
 	 */
-	public Date getBuyDate() {
-		return this.buyDate;
+	public void setSellDate(Date sellDate) {
+		this.sellDate = sellDate;
 	}
 
 	/**
-	 * 비 고을 설정합니다..
+	 * 판매/도태일자을 가져옵니다..
 	 * 
-	 * @param buyNote
-	 *            비 고
+	 * @return 판매/도태일자
 	 */
-	public void setBuyNote(String buyNote) {
-		this.buyNote = buyNote;
+	public Date getSellDate() {
+		return this.sellDate;
 	}
 
 	/**
-	 * 비 고을 가져옵니다..
+	 * 비고을 설정합니다..
 	 * 
-	 * @return 비 고
+	 * @param sellNote
+	 *            비고
 	 */
-	public String getBuyNote() {
-		return this.buyNote;
+	public void setSellNote(String sellNote) {
+		this.sellNote = sellNote;
 	}
 
 	/**
-	 * 구입 연락처을 설정합니다..
+	 * 비고을 가져옵니다..
 	 * 
-	 * @param buyPhoneNumber
-	 *            구입 연락처
+	 * @return 비고
 	 */
-	public void setBuyPhoneNumber(String buyPhoneNumber) {
-		this.buyPhoneNumber = buyPhoneNumber;
+	public String getSellNote() {
+		return this.sellNote;
 	}
 
 	/**
-	 * 구입 연락처을 가져옵니다..
+	 * 판매 연락처을 설정합니다..
 	 * 
-	 * @return 구입 연락처
+	 * @param sellPhoneNumber
+	 *            판매 연락처
 	 */
-	public String getBuyPhoneNumber() {
-		return this.buyPhoneNumber;
+	public void setSellPhoneNumber(String sellPhoneNumber) {
+		this.sellPhoneNumber = sellPhoneNumber;
+	}
+
+	/**
+	 * 판매 연락처을 가져옵니다..
+	 * 
+	 * @return 판매 연락처
+	 */
+	public String getSellPhoneNumber() {
+		return this.sellPhoneNumber;
 	}
 
 	/**
@@ -153,6 +143,14 @@ public class HmCattleBuyInfo implements Serializable {
 	 */
 	public HmCattleRegister getHmCattleRegister() {
 		return this.hmCattleRegister;
+	}
+
+	public String getEntityDiscernNo() {
+		return entityDiscernNo;
+	}
+
+	public void setEntityDiscernNo(String entityDiscernNo) {
+		this.entityDiscernNo = entityDiscernNo;
 	}
 
 	/**
@@ -180,7 +178,7 @@ public class HmCattleBuyInfo implements Serializable {
 		if (getClass() != obj.getClass()) {
 			return false;
 		}
-		HmCattleBuyInfo other = (HmCattleBuyInfo) obj;
+		HmCattleSellStoreInfo other = (HmCattleSellStoreInfo) obj;
 		if (hmCattleRegister == null) {
 			if (other.hmCattleRegister != null) {
 				return false;
