+++ skan.potal.com/src/main/java/com/skan/potal/web/potal/cattle/repository/CattleCalfRecodeRepository.java
@@ -20,8 +20,8 @@ package com.skan.potal.web.potal.cattle.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.skan.potal.web.potal.cattle.model.HmCattleCalfRecode;
-import com.skan.potal.web.potal.cattle.model.HmCattleCureInfoId;
+import com.skan.potal.web.potal.cattle.dto.HmCattleCalfRecode;
+import com.skan.potal.web.potal.cattle.dto.HmCattleCureInfoId;
 
 /**
  * @author ahn
